pyproject.toml
[project]
name = "dsrna-worst-case-pipeline-v2"
version = "0.1.0"
description = "Pipeline to fetch OrthoDB CDS for siRNA worst-case analysis"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "biopython",
    "httpx",
    "loguru",
    "matplotlib",
    "muscle>=0.0.4",
    "muscle3>=0.8.0",
    "numpy>=2.4.2",
    "pandas",
    "pyfamsa>=0.6.2",
    "pymsaviz>=0.5.0",
    "requests",
    "seaborn",
    "tqdm",
    "typer",
    "viennarna>=2.7.2",
]

[project.scripts]
dsrna-pipeline = "dsrna_worst_case_pipeline_v2.main:app"

[tool.uv]
package = true
